# packages/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StorefrontOut": {
        "properties": {
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "holiday_mode": {
            "title": "Holiday Mode",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "return_policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Policy"
          },
          "tagline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tagline"
          },
          "vendor_id": {
            "format": "uuid",
            "title": "Vendor Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "vendor_id",
          "tagline",
          "bio",
          "return_policy",
          "holiday_mode"
        ],
        "title": "StorefrontOut",
        "type": "object"
      },
      "StorefrontUpdate": {
        "properties": {
          "bio": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "holiday_mode": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Holiday Mode"
          },
          "return_policy": {
            "anyOf": [
              {
                "maxLength": 8000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Policy"
          },
          "tagline": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tagline"
          }
        },
        "title": "StorefrontUpdate",
        "type": "object"
      },
      "TenantCreate": {
        "properties": {
          "default_locale": {
            "default": "bn",
            "enum": [
              "bn",
              "en"
            ],
            "title": "Default Locale",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "owner_email": {
            "format": "email",
            "title": "Owner Email",
            "type": "string"
          },
          "owner_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Name"
          },
          "owner_password": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Password"
          },
          "slug": {
            "pattern": "^[a-z0-9](?:[a-z0-9-]{1,38}[a-z0-9])$",
            "title": "Slug",
            "type": "string"
          },
          "store_mode": {
            "default": "single",
            "enum": [
              "single",
              "multi"
            ],
            "title": "Store Mode",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "owner_email"
        ],
        "title": "TenantCreate",
        "type": "object"
      },
      "TenantCreated": {
        "properties": {
          "default_locale": {
            "title": "Default Locale",
            "type": "string"
          },
          "house_vendor_id": {
            "format": "uuid",
            "title": "House Vendor Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_user_id": {
            "format": "uuid",
            "title": "Owner User Id",
            "type": "string"
          },
          "primary_host": {
            "title": "Primary Host",
            "type": "string"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "store_mode": {
            "title": "Store Mode",
            "type": "string"
          }
        },
        "required": [
          "id",
          "public_id",
          "slug",
          "name",
          "status",
          "store_mode",
          "default_locale",
          "primary_host",
          "house_vendor_id",
          "owner_user_id"
        ],
        "title": "TenantCreated",
        "type": "object"
      },
      "TenantOut": {
        "properties": {
          "default_locale": {
            "title": "Default Locale",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "store_mode": {
            "title": "Store Mode",
            "type": "string"
          }
        },
        "required": [
          "id",
          "public_id",
          "slug",
          "name",
          "status",
          "store_mode",
          "default_locale"
        ],
        "title": "TenantOut",
        "type": "object"
      },
      "TenantPatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "trial",
                  "active",
                  "past_due",
                  "suspended",
                  "cancelled"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "store_mode": {
            "anyOf": [
              {
                "enum": [
                  "single",
                  "multi"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Mode"
          }
        },
        "title": "TenantPatch",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VendorCreate": {
        "properties": {
          "display_name": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Display Name",
            "type": "string"
          },
          "owner_email": {
            "format": "email",
            "title": "Owner Email",
            "type": "string"
          },
          "slug": {
            "pattern": "^[a-z0-9](?:[a-z0-9-]{1,38}[a-z0-9])$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "display_name",
          "owner_email"
        ],
        "title": "VendorCreate",
        "type": "object"
      },
      "VendorOut": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_house": {
            "title": "Is House",
            "type": "boolean"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "display_name",
          "status",
          "is_house"
        ],
        "title": "VendorOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ecomm SaaS marketplace API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/storefronts/{storefront_id}": {
      "get": {
        "operationId": "admin_get_storefront_api_v1_admin_storefronts__storefront_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "storefront_id",
            "required": true,
            "schema": {
              "title": "Storefront Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorefrontOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Get Storefront",
        "tags": [
          "admin:vendors"
        ]
      }
    },
    "/api/v1/admin/vendors": {
      "get": {
        "operationId": "list_vendors_api_v1_admin_vendors_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VendorOut"
                  },
                  "title": "Response List Vendors Api V1 Admin Vendors Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Vendors",
        "tags": [
          "admin:vendors"
        ]
      },
      "post": {
        "operationId": "create_vendor_api_v1_admin_vendors_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Vendor",
        "tags": [
          "admin:vendors"
        ]
      }
    },
    "/api/v1/admin/vendors/{vendor_id}": {
      "get": {
        "operationId": "get_vendor_api_v1_admin_vendors__vendor_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "vendor_id",
            "required": true,
            "schema": {
              "title": "Vendor Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Vendor",
        "tags": [
          "admin:vendors"
        ]
      }
    },
...
    "/api/v1/vendor/storefronts/{storefront_id}": {
      "get": {
        "operationId": "get_storefront_api_v1_vendor_storefronts__storefront_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "storefront_id",
            "required": true,
            "schema": {
              "title": "Storefront Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorefrontOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Storefront",
        "tags": [
          "vendor"
        ]
      },
      "patch": {
        "operationId": "update_storefront_api_v1_vendor_storefronts__storefront_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "storefront_id",
            "required": true,
            "schema": {
              "title": "Storefront Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorefrontUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorefrontOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Storefront",
        "tags": [
          "vendor"
        ]
      }
    },
...
    "/platform/v1/tenants": {
      "get": {
        "operationId": "list_tenants_platform_v1_tenants_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TenantOut"
                  },
                  "title": "Response List Tenants Platform V1 Tenants Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Tenants",
        "tags": [
          "platform"
        ]
      },
      "post": {
        "operationId": "create_tenant_platform_v1_tenants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Tenant",
        "tags": [
          "platform"
        ]
      }
    },
    "/platform/v1/tenants/{tenant_id}": {
      "patch": {
        "operationId": "patch_tenant_platform_v1_tenants__tenant_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Tenant",
        "tags": [
          "platform"
        ]
      }
    },
...
  }
}